Refactor: Move partydonation utils into own file
